Move db/api models to models.py and move disk models to disk.py
